Patches for issue #23, world loading bug
Former-commit-id: f10f66adc1a2818993d8259f4c24819b9588fa4d
Former-commit-id: b415310cc09afb0d5679e027fecf10fb6b2f2cdf

+++ Source/EdenUniverseBuilder/WorldConverter.h
@@ -28,6 +28,8 @@ protected:
 
 public:
 
+	//void ReloadChunks();
+
 	typedef struct blockData
 	{
 
@@ -109,7 +111,7 @@ public:
 	USceneComponent* Root;
 
 	UPROPERTY(EditAnywhere)
-	int render_distance = 2;
+	float render_distance = 2.0;
 
 	UPROPERTY(VisibleAnywhere)
 	class UInstancedStaticMeshComponent* Mesh;
